Fix duplicate authentication
Summary:
The response to EHLO triggers an authentication command, but with TLS
two EHLOs are sent: For the 220 from the server and after TLS negotiation.
However, sending it twice results in an unexpected "503 already authenticated"
response which ends up getting parsed by the SendJob, causing confusion.

By leaving the EHLO-resending to the SessionPrivate, the state can be properly
tracked.

BUG: 387926
BUG: 388068

Reviewers: mlaurent, dvratil

Subscribers: lbeltrame, cgiboudeaux

Differential Revision: https://phabricator.kde.org/D9476