KSMTP: improve error handling.
Summary:
Sending a mail to doesnotexist@example.com returns a 450 error (at least
on a Kolab server), which was wrongly displayed as "Server time out" to the
user (e.g. above the mail header, in the kmail outbox).

Now this testcase displays
"Failed to transport message. Mailbox unavailable. The server said: 4.1.1 <doesnotexit@example.com>: Recipient address rejected: User unknown in local recipient table"

It's slightly redundant, but the point is that "Mailbox unavailable" can
be translated, while the error text from the server will likely not be.

Reviewers: dvratil, lbeltrame, mlaurent, fvogt

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D15528