Connect QSslSocket::encrypted only once
We can enter startSsl() multiple times, when a socket is reconnected in
Session, which then would result in multiple connections. Connecting it
earlier OTOH shouldn't hurt, so connect it together with the rest of the
socket signals after construction.

We'll also need this for starting encryption earlier than in LoginJob.